+++ packages/go-kosu/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+-   Ethereum Snapshot block
+
 ## v0.3.0
 
 -   introduce validators verification from app_state